requirements.txt
flask==3.1.3
flask-cors==6.0.2
gspread==6.2.1
google-auth==2.49.1
